Azure Bot/requirements.txt
langchain 
langchain-openai
streamlit
python-dotenv
langchain_community